requirements.txt
Werkzeug==2.0.1
Flask==2.0.1
Flask_SQLAlchemy==2.5.1
Pillow==8.2.0
